refactor(models): break models→commands import cycle (#576)
Move lock-file persistence and orphan-cleanup helpers out of
commands/_sync_helpers.py into a new models-layer module
models/_git/lock_io.py, where the 3-way merge strategy actually drives
them. merge.py previously reached up into the commands layer via
function-local imports (a cycle hidden behind deferred imports, since a
module-level import would deadlock at load time).

- New models/_git/lock_io.py owns LOCK_FILE, _write_lock,
  _lock_content_unchanged, _read_previously_tracked_files,
  _clean_orphaned_files, _delete_orphaned_file, _files_from_snapshot,
  _warn_about_workflow_files, and the fcntl/_FCNTL_AVAILABLE block.
- merge.py now imports lock_io at module level and calls it as
  lock_io.<fn>(...); the two function-local imports are removed.
- _sync_helpers.py is reduced to a re-export shim: it keeps the
  command-only _load_lock_or_warn and re-exports the moved names so
  existing `from rhiza.commands._sync_helpers import ...` call sites
  (tree.py, status.py, tests) keep working. Dependency direction is now
  commands -> models.
- Tests: patch targets for the moved symbols point at
  rhiza.models._git.lock_io.

No models -> commands import remains; all imports are module-level.
Gates green: fmt, typecheck (ty + mypy --strict, 35 files),
test (649 passed, 100% coverage).

Closes #576
